\name{eolsonde-package}
\alias{eolsonde-package}
\alias{eolsonde}
\docType{package}
\title{
Read and Visualize EOL sounding data
}
\description{
This package contains functions for reading EOL sounding files
    and plotting the soundings.
}
\details{
\tabular{ll}{
Package: \tab eolsonde\cr
Version: \tab 2.4-268\cr
Depends: \tab lattice,maps, eolts, isfs\cr
License: \tab BSD_2_clause + file LICENSE\cr
Date: \tab 2014-05-01\cr
Packaged: \tab 2014-05-20 19:33:44 UTC; maclean\cr
Built: \tab R 3.0.2; ; 2014-05-20 19:33:45 UTC; unix\cr
}

Index:
\preformatted{
eolsonde        Read and Visualize EOL sounding data
readDFile       Read a D file.
readQCFile      Read a QC file.
readSoundings   Read one or more D or QC files.
sprofile        Plot profiles one or more variables in one sounding.
svprofile       Plot profiles of one variable over multiple soundings
scontour        Plot contours of a variable over X=sounding, Y=pressure or altitude
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Liangying Zhang
Maintainer: Gordon Maclean <maclean@ucar.edu>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[eolts:eolts-package]{eolts}}
\code{\link[isfs:isfs-package]{isfs}}
}
\examples{
}
